% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports_and_presentations.R
\name{create_bioinfo_revealjs}
\alias{create_bioinfo_revealjs}
\title{Create a Quarto presentation, Bioinfo-group styled}
\usage{
create_bioinfo_revealjs(
  file_name = NULL,
  presentation_folder = NULL,
  ext_name = "bioinfo-revealjs",
  expose_aux_files = TRUE
)
}
\arguments{
\item{file_name}{Character string, the name of the file for the report -
no need to specify the qmd extension.
If preparing a presentation that will end up getting published online, it might
be clever to rename the rendered file into \code{index.html} (or start already
having an index.qmd file, but in the end it is a good practice to keep
informative file names)}

\item{presentation_folder}{Character string, pointing to the folder where the
presentation and its associated files should be created. Defaults to NULL, which
falls back to the current working directory.}

\item{ext_name}{Character string, defaulting to "bioinfo-revealjs". Other values
(possibly in other functions) would be "bioinfo-html" or similar...}

\item{expose_aux_files}{Logical value, whether to expose&copy the auxiliary
files (scss, and possibly a .bib file as well) into the toplevel of the
specified folder.}
}
\value{
Invisibly, the path to the created quarto source file for the
presentation
}
\description{
Create a Quarto presentation, Bioinfo-group styled
}
\examples{
temp_folder <- tempdir()
create_bioinfo_revealjs(file_name = "mypresentation",
                        presentation_folder = temp_folder)
}
\seealso{
Other quartodocuments: 
\code{\link{create_bioinfo_html}()}
}
\concept{quartodocuments}
