% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_IMBEI_beamer.R
\name{create_IMBEI_beamer}
\alias{create_IMBEI_beamer}
\title{Create an R Markdown Beamer presentation, IMBEI styled}
\usage{
create_IMBEI_beamer(file_name = NULL, folder_name = NULL)
}
\arguments{
\item{file_name}{Character string, the name of the file for the presentation -
no need to specify the .Rmd extension.}

\item{folder_name}{Character string, pointing to the folder where the
presentation and its associated files should be created. Defaults to NULL, which
falls back to the current working directory.}
}
\value{
Invisibly, the path to the created R Markdown source file for the
presentation
}
\description{
Create an R Markdown Beamer presentation, IMBEI styled
}
