% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMBEIstyle-pkg.R
\docType{package}
\name{IMBEIstyle-pkg}
\alias{IMBEIstyle}
\alias{IMBEIstyle-package}
\alias{IMBEIstyle-pkg}
\title{IMBEIstyle}
\description{
This is a collection of functions (and templates, and style sheets) to write reports and create presentations, in the IMBEI style.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/imbeimainz/IMBEIstyle}
  \item Report bugs at \url{https://github.com/imbeimainz/IMBEIstyle/issues}
}

}
\author{
\strong{Maintainer}: Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})

}
\keyword{internal}
